+++ packages/lib/src/dayjs/date-formatter.ts
@@ -1,4 +1,5 @@
 import { dayjs } from "./index";
+import { CURRENT_TIMEZONE } from "./time-zone.const";
 
 type DateInput = Date | string | number | null | undefined;
 
@@ -11,16 +12,10 @@ export interface DateFormatterOptions {
 }
 
 const DEFAULT_LOCALE: AppLocale = "en-US";
-const DEFAULT_TIME_ZONE = "UTC";
 
 export const getBrowserLocale = () => {
   if (typeof navigator === "undefined") return DEFAULT_LOCALE;
   return navigator.language || DEFAULT_LOCALE;
-};
-
-export const getBrowserTimeZone = () => {
-  if (typeof Intl === "undefined") return DEFAULT_TIME_ZONE;
-  return Intl.DateTimeFormat().resolvedOptions().timeZone || DEFAULT_TIME_ZONE;
 };
 
 export const normalizeLocale = (locale?: string): AppLocale => {
@@ -47,7 +42,7 @@ export const resolveDateFormatterOptions = (
 ) => {
   return {
     locale: normalizeLocale(options.locale ?? getBrowserLocale()),
-    timeZone: options.timeZone ?? getBrowserTimeZone(),
+    timeZone: options.timeZone ?? CURRENT_TIMEZONE,
     timeFormat: options.timeFormat ?? 24,
   };
 };
@@ -55,6 +50,48 @@ export const resolveDateFormatterOptions = (
 export type ResolvedDateFormatterOptions = ReturnType<
   typeof resolveDateFormatterOptions
 >;
+
+export const formatDate = (
+  date: DateInput,
+  format = "YYYY-MM-DD",
+  options: DateFormatterOptions = {},
+): string => {
+  if (!date) return "";
+
+  const { locale, timeZone } = resolveDateFormatterOptions(options);
+  const target = dayjs(date).tz(timeZone);
+  if (!target.isValid()) return "";
+
+  return target.locale(toDayjsLocale(locale)).format(format);
+};
+
+export const formatTime = (
+  date: DateInput,
+  options: DateFormatterOptions = {},
+): string => {
+  if (!date) return "";
+
+  const { timeZone, timeFormat } = resolveDateFormatterOptions(options);
+  const target = dayjs(date).tz(timeZone);
+  if (!target.isValid()) return "";
+
+  return target.format(timeFormat === 12 ? "h:mma" : "HH:mm");
+};
+
+export const formatTimeRange = (
+  startedAt: DateInput,
+  closedAt: DateInput,
+  options: DateFormatterOptions = {},
+): string => {
+  if (!startedAt || !closedAt) return "";
+
+  const start = formatTime(startedAt, options);
+  const end = formatTime(closedAt, options);
+
+  if (!start || !end) return "";
+
+  return `${start} - ${end}`;
+};
 
 /** 고정 일요일(ISO). `weekdayIndex`만큼 더해 요일 이름을 만들 때 사용한다. */
 const SUNDAY_REF = "2024-01-07";
@@ -73,10 +110,6 @@ export const formatWeekdayShort = (
     .locale(toDayjsLocale(locale))
     .format("dd");
 };
-
-/** 시간대 라벨 등에 쓰는 2자리 문자열 (예: 3 → "03"). */
-export const formatTwoDigitNumber = (n: number): string =>
-  String(Math.trunc(n)).padStart(2, "0");
 
 /**
  * 지역별 날짜/시간 포맷.
@@ -108,83 +141,6 @@ export const formatLocalizedDateTime = (
   }).format(target);
 };
 
-export const formatDate = (
-  date: DateInput,
-  format = "YYYY-MM-DD",
-  options: DateFormatterOptions = {},
-): string => {
-  if (!date) return "";
-
-  const { locale, timeZone } = resolveDateFormatterOptions(options);
-
-  const target = dayjs(date).tz(timeZone);
-
-  if (!target.isValid()) return "";
-
-  return target.locale(toDayjsLocale(locale)).format(format);
-};
-
-export const formatTime = (
-  date: DateInput,
-  options: DateFormatterOptions = {},
-): string => {
-  if (!date) return "";
-
-  const { timeZone, timeFormat } = resolveDateFormatterOptions(options);
-
-  const target = dayjs(date).tz(timeZone);
-
-  if (!target.isValid()) return "";
-
-  return target.format(timeFormat === 12 ? "h:mma" : "HH:mm");
-};
-
-export const formatDuration = (
-  seconds: number,
-  t: (key: string, options?: Record<string, unknown>) => string,
-): string => {
-  if (seconds <= 0 || Number.isNaN(seconds)) {
-    return t("common:duration.second", { count: 0 });
-  }
-
-  const totalSeconds = Math.floor(seconds);
-
-  const hours = Math.floor(totalSeconds / 3600);
-  const minutes = Math.floor((totalSeconds % 3600) / 60);
-  const secs = totalSeconds % 60;
-
-  const parts: string[] = [];
-
-  if (hours > 0) {
-    parts.push(t("common:duration.hour", { count: hours }));
-  }
-
-  if (minutes > 0) {
-    parts.push(t("common:duration.minute", { count: minutes }));
-  }
-
-  if (secs > 0 || parts.length === 0) {
-    parts.push(t("common:duration.second", { count: secs }));
-  }
-
-  return parts.slice(0, 2).join(" ");
-};
-
-export const formatTimeRange = (
-  startedAt: DateInput,
-  closedAt: DateInput,
-  options: DateFormatterOptions = {},
-): string => {
-  if (!startedAt || !closedAt) return "";
-
-  const start = formatTime(startedAt, options);
-  const end = formatTime(closedAt, options);
-
-  if (!start || !end) return "";
-
-  return `${start} - ${end}`;
-};
-
 /**
  * Unix timestamp(초) 두 개의 차이를 초 단위 실수로 반환.
  *
@@ -203,13 +159,3 @@ export const calculateTimeDiff = (
 
   return closed.diff(visited, "second", true);
 };
-
-export const DATE_FORMAT = {
-  YYYY_MM_DD: "YYYY.MM.DD",
-  YYYY_MM_DD_DASH: "YYYY-MM-DD",
-  YYYY_MM_DD_SLASH: "YYYY/MM/DD",
-  MM_DD: "MM.DD",
-  YYYY_MM: "YYYY.MM",
-} as const;
-
-export type DateFormat = (typeof DATE_FORMAT)[keyof typeof DATE_FORMAT];
