+++ packages/lib/src/dayjs/index.ts
@@ -14,4 +14,8 @@ dayjs.extend(timezone);
 dayjs.extend(localizedFormat);
 
 export { dayjs };
+export * from "./date-format.const";
+export * from "./date-formatter";
+export * from "./duration-formatter";
+export * from "./time-zone.const";
 export type { Dayjs } from "dayjs";
